refactor(core): split dag-executor.ts into focused modules
Split the 1482-line dag-executor.ts into 6 focused files:

- dag-executor.ts (740 lines) — executeAgentDag, topological sort,
  DagExecutorDeps/DagExecuteOptions types, run lifecycle
- node-env.ts (179 lines) — buildNodeEnv, mergedInputs,
  buildUpstreamSnapshot, filterEnvForLog, env allowlists, secret
  redaction patterns
- node-templates.ts (41 lines) — resolveUpstreamTemplate,
  resolveVarsTemplate
- node-spawner.ts (126 lines) — spawnNodeReal, spawnProcess,
  SpawnResult/SpawnNodeFn types, prompt template resolution
- flow-control.ts (341 lines) — executeControlFlowNode, conditional,
  switch, agent-invoke, loop, evaluateOnlyIf, walkPath
- tool-dispatch.ts (39 lines) — resolveToolId, resolveToolInputs

Zero behavior change. All 561 tests pass unchanged. Public API preserved
via re-exports from dag-executor.ts (resolveUpstreamTemplate,
resolveVarsTemplate, SpawnResult, SpawnNodeFn).

Circular dependency between dag-executor and flow-control avoided by
passing executeAgentDag as a callback (ExecuteDagFn) to agent-invoke
and loop node executors.
