+++ packages/core/src/dag-executor.ts
@@ -20,24 +20,30 @@
  */
 
 import { randomUUID } from 'node:crypto';
-import type { ChildProcess } from 'node:child_process';
-import { spawn } from 'node:child_process';
-import type { Agent, AgentNode, NodeErrorCategory, NodeOutput, NodeStructuredOutput, OnlyIfCondition } from './agent-v2-types.js';
+import type { Agent, AgentNode, NodeErrorCategory, NodeOutput, NodeStructuredOutput } from './agent-v2-types.js';
 import type { Run, RunStatus } from './types.js';
 import type { RunStore } from './run-store.js';
 import type { AgentStore } from './agent-store.js';
 import type { SecretsStore } from './secrets-store.js';
 import type { ToolStore } from './tool-store.js';
-import { type VariablesStore, looksLikeSensitive } from './variables-store.js';
+import type { VariablesStore } from './variables-store.js';
 import type { ToolOutput, BuiltinToolContext } from './tool-types.js';
 import { getBuiltinTool } from './builtin-tools.js';
 import { buildToolOutput } from './output-framing.js';
+import { UntrustedCommunityShellError } from './agent-executor.js';
+import { buildNodeEnv, buildUpstreamSnapshot, filterEnvForLog } from './node-env.js';
+import { type SpawnResult, type SpawnNodeFn, spawnNodeReal } from './node-spawner.js';
+import { resolveToolId, resolveToolInputs } from './tool-dispatch.js';
 import {
-  UntrustedCommunityShellError,
-  type ExecutionResult,
-} from './agent-executor.js';
-import { substituteInputs, SENSITIVE_ENV_NAMES } from './input-resolver.js';
-import { extractUpstreamReferences } from './agent-v2-schema.js';
+  executeControlFlowNode,
+  executeAgentInvokeNode,
+  executeLoopNode,
+  evaluateOnlyIf,
+} from './flow-control.js';
+
+// Re-export for consumers that import from dag-executor.
+export { resolveUpstreamTemplate, resolveVarsTemplate } from './node-templates.js';
+export type { SpawnResult, SpawnNodeFn } from './node-spawner.js';
 
 export interface DagExecutorDeps {
   runStore: RunStore;
@@ -113,18 +119,6 @@ export interface DagExecuteOptions {
   parentNodeId?: string;
 }
 
-/**
- * Returned by the real spawner; also the shape test doubles return.
- * Matches `ExecutionResult` from agent-executor.ts (duplicated here to
- * keep the DAG executor decoupled from the v1 agent-level shape).
- */
-type SpawnResult = ExecutionResult & { category?: NodeErrorCategory };
-
-type SpawnNodeFn = (
-  node: AgentNode,
-  env: Record<string, string>,
-  opts: { agentId: string; agentSource: Agent['source']; allowUntrustedShell?: ReadonlySet<string> },
-) => Promise<SpawnResult>;
 
 /**
  * Execute an agent's DAG end-to-end. Creates the parent `runs` row,
@@ -389,7 +383,7 @@ export async function executeAgentDag(
     }
 
     if (node.type === 'loop') {
-      const loopResult = await executeLoopNode(node, outputs, runId, options, deps, agent);
+      const loopResult = await executeLoopNode(node, outputs, runId, options, deps, agent, executeAgentDag);
       const completedAt = new Date().toISOString();
       if (loopResult.ok) {
         const outputsJson = JSON.stringify(loopResult.output);
@@ -427,7 +421,7 @@ export async function executeAgentDag(
     }
 
     if (node.type === 'agent-invoke') {
-      const invokeResult = await executeAgentInvokeNode(node, outputs, runId, options, deps, agent);
+      const invokeResult = await executeAgentInvokeNode(node, outputs, runId, options, deps, agent, executeAgentDag);
       const completedAt = new Date().toISOString();
       if (invokeResult.ok) {
         const outputsJson = invokeResult.output ? JSON.stringify(invokeResult.output) : undefined;
@@ -743,740 +737,4 @@ export function topologicalSort(nodes: AgentNode[]): AgentNode[] {
     throw new Error(`Cycle detected in DAG (schema should have rejected this)`);
   }
   return order;
-}
-
-// -- Environment build --
-
-/**
- * Build the env map a node will spawn with. Layers, in order of precedence:
- *
- *   1. MINIMAL_ALLOWLIST / LOCAL_ALLOWLIST from current process.env
- *      (same as v1 per-trust env filter; community agents get MINIMAL)
- *   2. Node's `envAllowlist` adds
- *   3. Node's `env:` YAML field (values with {{inputs.X}} substituted)
- *   4. Node's declared `secrets:` from secretsStore
- *   5. Resolved agent inputs as env vars (scrubbed of sensitive names)
- *   6. `UPSTREAM_<NODEID>_RESULT` for each declared upstream
- *
- * Later layers win on key collisions. Sensitive names (PATH, NODE_OPTIONS,
- * LD_PRELOAD, etc.) coming from user-supplied inputs are blocked — the
- * schema rejects these at load time but we defense-in-depth here too.
- */
-async function buildNodeEnv(
-  agent: Agent,
-  node: AgentNode,
-  callerInputs: Record<string, string>,
-  upstreamSnapshot: Record<string, string>,
-  deps: DagExecutorDeps,
-): Promise<Record<string, string>> {
-  const trustLevel = agent.source === 'community' ? 'community' : 'local';
-  const baseAllowlist = trustLevel === 'community' ? MINIMAL_ALLOWLIST : LOCAL_ALLOWLIST;
-  const allowed = new Set<string>([...baseAllowlist, ...(node.envAllowlist ?? [])]);
-  const env: Record<string, string> = {};
-
-  // 1 + 2: process.env filtered by allowlist.
-  for (const [k, v] of Object.entries(process.env)) {
-    if (v === undefined) continue;
-    if (allowed.has(k)) env[k] = v;
-    else if (trustLevel === 'local' && LOCAL_PATTERNS.some((re) => re.test(k))) env[k] = v;
-  }
-
-  // 3: node's YAML env: — templates substituted.
-  if (node.env) {
-    for (const [k, v] of Object.entries(node.env)) {
-      let resolved = resolveUpstreamTemplate(v, upstreamSnapshot);
-      if (deps.variablesStore) resolved = resolveVarsTemplate(resolved, deps.variablesStore.getAll());
-      env[k] = substituteInputs(resolved, mergedInputs(agent, callerInputs));
-    }
-  }
-
-  // 4: secrets. Each node only gets what it declares; never shares across nodes.
-  if (node.secrets && node.secrets.length > 0) {
-    if (!deps.secretsStore) {
-      throw new Error(
-        `Node "${node.id}" declares secrets but no secrets store is available. ` +
-        `Pass one via DagExecutorDeps.secretsStore.`,
-      );
-    }
-    const all = await deps.secretsStore.getAll();
-    const missing: string[] = [];
-    for (const name of node.secrets) {
-      if (name in all) env[name] = all[name];
-      else missing.push(name);
-    }
-    if (missing.length > 0) {
-      throw new Error(`Missing secrets for node "${node.id}": ${missing.join(', ')}. Run 'sua secrets set <name>'.`);
-    }
-  }
-
-  // 5: global variables. Injected as $NAME for shell nodes. Lower
-  // precedence than inputs (step 6) — inputs override variables.
-  if (deps.variablesStore) {
-    for (const [k, v] of Object.entries(deps.variablesStore.getAll())) {
-      if (SENSITIVE_ENV_NAMES.has(k)) continue;
-      env[k] = v;
-    }
-  }
-
-  // 6: caller-supplied inputs. Drop sensitive names as a belt-and-suspenders
-  // on the schema check. Merge agent-level defaults for missing values.
-  // These override global variables (step 5) by being applied after.
-  for (const [k, v] of Object.entries(mergedInputs(agent, callerInputs))) {
-    if (SENSITIVE_ENV_NAMES.has(k)) continue;
-    env[k] = v;
-  }
-
-  // 7: upstream results as UPSTREAM_<NODEID>_RESULT.
-  for (const [upstreamId, value] of Object.entries(upstreamSnapshot)) {
-    const key = `UPSTREAM_${upstreamId.toUpperCase().replace(/-/g, '_')}_RESULT`;
-    env[key] = value;
-  }
-
-  return env;
-}
-
-/**
- * Compose caller-supplied inputs with agent-level defaults. Uses the
- * existing typed-input machinery via a narrow helper — missing required
- * inputs throw, bad types throw, unknown keys are dropped (the agent
- * declared what it wants; extras are the caller's problem, not the node's).
- */
-function mergedInputs(agent: Agent, callerInputs: Record<string, string>): Record<string, string> {
-  const out: Record<string, string> = {};
-  const specs = agent.inputs ?? {};
-  const declared = new Set(Object.keys(specs));
-
-  for (const [name, spec] of Object.entries(specs)) {
-    if (name in callerInputs) {
-      out[name] = callerInputs[name];
-    } else if (spec.default !== undefined) {
-      out[name] = String(spec.default);
-    } else if (spec.required !== false) {
-      throw new Error(`Missing required input "${name}" for agent "${agent.id}"`);
-    }
-  }
-
-  // Undeclared caller inputs are dropped — they can't reach a node. Schema
-  // enforced this at load time for run.ts; the executor is lenient here
-  // because scheduled / MCP invocations pass caller-wide input bags.
-  for (const [k, v] of Object.entries(callerInputs)) {
-    if (!declared.has(k)) continue;
-    if (!(k in out)) out[k] = v;
-  }
-
-  return out;
-}
-
-/**
- * Snapshot of upstream node results that feed this node. Built from the
- * `outputs` map after execution of earlier nodes.
- */
-function buildUpstreamSnapshot(node: AgentNode, outputs: Map<string, NodeOutput>): Record<string, string> {
-  const out: Record<string, string> = {};
-  for (const dep of node.dependsOn ?? []) {
-    const o = outputs.get(dep);
-    if (o) out[dep] = o.result;
-  }
-  return out;
-}
-
-/**
- * Substitute `{{upstream.<id>.result}}` tokens in a text blob. Deliberately
- * kept tiny and greedy — we rely on schema-time validation to guarantee
- * every reference resolves.
- */
-export function resolveUpstreamTemplate(text: string, snapshot: Record<string, string>): string {
-  if (!text.includes('{{upstream.')) return text;
-  const refs = extractUpstreamReferences(text);
-  let out = text;
-  for (const id of refs) {
-    const value = snapshot[id] ?? '';
-    // Escape {{ in the substituted value so the inputs resolver that runs
-    // afterwards can't re-expand it as a second template layer. Same
-    // defense the v1 chain-resolver ships (chain-resolver.ts:120-125).
-    const safe = value.replace(/\{\{/g, '{ {');
-    // Use a simple literal replace; the ref format is fixed.
-    out = out.split(`{{upstream.${id}.result}}`).join(safe);
-  }
-  return out;
-}
-
-/**
- * Substitute `{{vars.<NAME>}}` tokens from the global variables store.
- * Runs after upstream resolution but before input substitution.
- */
-export function resolveVarsTemplate(text: string, vars: Record<string, string>): string {
-  if (!text.includes('{{vars.')) return text;
-  let out = text;
-  for (const [name, value] of Object.entries(vars)) {
-    const safe = value.replace(/\{\{/g, '{ {');
-    out = out.split(`{{vars.${name}}}`).join(safe);
-  }
-  return out;
-}
-
-/**
- * Patterns that look like secret values even when the env-var name
- * doesn't match the sensitive-name heuristic. Catches common token
- * formats that leak through env inheritance or non-standard naming.
- */
-const SENSITIVE_VALUE_PATTERNS = [
-  /^ghp_[A-Za-z0-9]{36,}$/,      // GitHub PATs
-  /^gho_[A-Za-z0-9]{36,}$/,      // GitHub OAuth tokens
-  /^sk-[A-Za-z0-9]{20,}$/,       // OpenAI / Stripe secret keys
-  /^xox[bpars]-[A-Za-z0-9-]+$/,  // Slack tokens
-  /^AKIA[A-Z0-9]{16}$/,          // AWS access key IDs
-  /^eyJ[A-Za-z0-9_-]{20,}\.[A-Za-z0-9_-]{20,}/, // JWTs
-];
-
-/**
- * Redact secret values from the env map before we serialise it to
- * `inputs_json` for the node_executions row. Three layers:
- *
- *   1. Declared secrets — names in the node's `secrets:` array.
- *   2. Sensitive names — names matching TOKEN, KEY, SECRET, PASS, etc.
- *      (catches variables the user forgot to declare as secrets).
- *   3. Sensitive values — values matching known credential formats
- *      (GitHub PATs, Slack tokens, AWS keys, JWTs, etc.).
- *
- * Prefer false-positive redaction over leaking credentials in run logs.
- */
-function filterEnvForLog(env: Record<string, string>, node: AgentNode): Record<string, string> {
-  const out: Record<string, string> = {};
-  const secrets = new Set(node.secrets ?? []);
-  for (const [k, v] of Object.entries(env)) {
-    if (
-      secrets.has(k) ||
-      looksLikeSensitive(k) ||
-      SENSITIVE_VALUE_PATTERNS.some((re) => re.test(v))
-    ) {
-      out[k] = '<redacted>';
-    } else {
-      out[k] = v;
-    }
-  }
-  return out;
-}
-
-// -- Real spawner (production path) --
-
-/**
- * Copy of the spawn machinery from agent-executor.ts, adapted for an
- * `AgentNode`. We don't go through `executeAgent` because it expects a v1
- * `AgentDefinition` and requires an adapter at the boundary; duplicating
- * ~40 lines of spawn plumbing is cheaper and keeps the DAG executor
- * decoupled from the v1 flow while v1 still exists.
- *
- * Categorisation notes:
- *   - exit code 124 = timeout we set (matches SIGTERM by convention)
- *   - exit code 127 = spawn failure (ENOENT / EACCES)
- *   - any other non-zero = exit_nonzero (returned as-is, executor categorises)
- */
-async function spawnNodeReal(
-  node: AgentNode,
-  env: Record<string, string>,
-  opts: { agentId: string; agentSource: Agent['source']; allowUntrustedShell?: ReadonlySet<string> },
-): Promise<SpawnResult> {
-  if (node.type === 'shell') {
-    if (!node.command) {
-      return { result: '', exitCode: 1, error: `Shell node "${node.id}" has no command`, category: 'setup' };
-    }
-    return spawnProcess('bash', ['-c', node.command], {
-      cwd: node.workingDirectory,
-      env,
-      timeoutSec: node.timeout ?? 300,
-    });
-  }
-
-  // claude-code — resolve {{inputs.X}}, {{upstream.X.result}}, {{vars.X}}
-  // in the prompt before passing to the CLI.
-  if (!node.prompt) {
-    return { result: '', exitCode: 1, error: `Claude-code node "${node.id}" has no prompt`, category: 'setup' };
-  }
-  let resolvedPrompt = node.prompt;
-  // Build upstream map from env's UPSTREAM_*_RESULT keys
-  const upstreamMap: Record<string, string> = {};
-  for (const [k, v] of Object.entries(env)) {
-    const m = k.match(/^UPSTREAM_(.+)_RESULT$/);
-    if (m) upstreamMap[m[1].toLowerCase().replace(/_/g, '-')] = v;
-  }
-  resolvedPrompt = resolveUpstreamTemplate(resolvedPrompt, upstreamMap);
-  resolvedPrompt = resolveVarsTemplate(resolvedPrompt, env);
-  resolvedPrompt = substituteInputs(resolvedPrompt, env);
-  const args = ['--print', resolvedPrompt];
-  if (node.model) { args.push('--model', node.model); }
-  if (node.maxTurns) { args.push('--max-turns', String(node.maxTurns)); }
-  if (node.allowedTools?.length) { args.push('--allowedTools', node.allowedTools.join(',')); }
-  return spawnProcess('claude', args, {
-    cwd: node.workingDirectory,
-    env,
-    timeoutSec: node.timeout ?? 300,
-  });
-}
-
-async function spawnProcess(
-  bin: string,
-  args: string[],
-  opts: { cwd?: string; env: Record<string, string>; timeoutSec: number },
-): Promise<SpawnResult> {
-  return new Promise<SpawnResult>((resolve) => {
-    let child: ChildProcess;
-    let killed = false;
-    try {
-      child = spawn(bin, args, {
-        cwd: opts.cwd,
-        env: opts.env,
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-    } catch (err) {
-      resolve({ result: '', exitCode: 127, error: (err as Error).message, category: 'spawn_failure' });
-      return;
-    }
-
-    let stdout = '';
-    let stderr = '';
-    child.stdout!.on('data', (d: Buffer) => { stdout += d.toString(); });
-    child.stderr!.on('data', (d: Buffer) => { stderr += d.toString(); });
-
-    const timer = setTimeout(() => {
-      killed = true;
-      child.kill('SIGTERM');
-      setTimeout(() => { if (!child.killed) child.kill('SIGKILL'); }, 5000);
-    }, opts.timeoutSec * 1000);
-
-    child.on('close', (code) => {
-      clearTimeout(timer);
-      if (killed) {
-        resolve({ result: stdout, exitCode: 124, error: `Timed out after ${opts.timeoutSec}s`, category: 'timeout' });
-      } else if (code === 0) {
-        resolve({ result: stdout, exitCode: 0 });
-      } else {
-        resolve({
-          result: stdout,
-          exitCode: code ?? 1,
-          error: stderr || `Process exited with code ${code}`,
-          category: 'exit_nonzero',
-        });
-      }
-    });
-
-    child.on('error', (err) => {
-      clearTimeout(timer);
-      resolve({ result: '', exitCode: 127, error: err.message, category: 'spawn_failure' });
-    });
-  });
-}
-
-// -- Env allowlists (duplicate of env-builder constants; keeping here lets
-//    the DAG executor build env without the v1 AgentDefinition intermediary.
-//    Extract to a shared module in a follow-up if this duplicates further.) --
-
-// -- Control-flow node dispatch --
-
-interface ControlFlowResult {
-  ok: boolean;
-  output: Record<string, unknown>;
-  error?: string;
-}
-
-/**
- * Execute a `conditional` or `switch` node. Pure in-process evaluation —
- * no child process, no env, no secrets. Reads the upstream structured
- * output and produces a result downstream nodes consume.
- */
-function executeControlFlowNode(
-  node: AgentNode,
-  outputs: Map<string, NodeOutput>,
-): ControlFlowResult {
-  if (node.type === 'conditional') {
-    return executeConditionalNode(node, outputs);
-  }
-  if (node.type === 'switch') {
-    return executeSwitchNode(node, outputs);
-  }
-  return { ok: false, output: {}, error: `Unknown control-flow type: ${node.type}` };
-}
-
-function executeConditionalNode(
-  node: AgentNode,
-  outputs: Map<string, NodeOutput>,
-): ControlFlowResult {
-  const config = node.conditionalConfig;
-  if (!config) {
-    return { ok: false, output: {}, error: 'Conditional node missing conditionalConfig' };
-  }
-  // The conditional evaluates against its first upstream's output.
-  const upstreamId = (node.dependsOn ?? [])[0];
-  if (!upstreamId) {
-    return { ok: false, output: {}, error: 'Conditional node has no dependsOn' };
-  }
-  const upOutput = outputs.get(upstreamId);
-  if (!upOutput) {
-    return { ok: false, output: {}, error: `Upstream "${upstreamId}" has no output` };
-  }
-
-  // Resolve the field value from structured output or parsed JSON result.
-  let value: unknown;
-  if (upOutput.outputs) {
-    value = walkPath(upOutput.outputs, config.predicate.field);
-  } else {
-    try {
-      const parsed = JSON.parse(upOutput.result);
-      value = walkPath(parsed, config.predicate.field);
-    } catch {
-      value = config.predicate.field === 'result' ? upOutput.result : undefined;
-    }
-  }
-
-  let matched = false;
-  if (config.predicate.exists !== undefined) {
-    matched = config.predicate.exists
-      ? (value !== undefined && value !== null)
-      : (value === undefined || value === null);
-  } else if (config.predicate.equals !== undefined) {
-    matched = value == config.predicate.equals;
-  } else if (config.predicate.notEquals !== undefined) {
-    matched = value != config.predicate.notEquals;
-  } else {
-    matched = value !== undefined && value !== null;
-  }
-
-  return { ok: true, output: { matched, value } };
-}
-
-function executeSwitchNode(
-  node: AgentNode,
-  outputs: Map<string, NodeOutput>,
-): ControlFlowResult {
-  const config = node.switchConfig;
-  if (!config) {
-    return { ok: false, output: {}, error: 'Switch node missing switchConfig' };
-  }
-  const upstreamId = (node.dependsOn ?? [])[0];
-  if (!upstreamId) {
-    return { ok: false, output: {}, error: 'Switch node has no dependsOn' };
-  }
-  const upOutput = outputs.get(upstreamId);
-  if (!upOutput) {
-    return { ok: false, output: {}, error: `Upstream "${upstreamId}" has no output` };
-  }
-
-  let value: unknown;
-  if (upOutput.outputs) {
-    value = walkPath(upOutput.outputs, config.field);
-  } else {
-    try {
-      const parsed = JSON.parse(upOutput.result);
-      value = walkPath(parsed, config.field);
-    } catch {
-      value = config.field === 'result' ? upOutput.result : undefined;
-    }
-  }
-
-  // Find matching case. Cases are keyed by name; the value stored is what
-  // the field should match. If none match, case = 'default' (always present
-  // implicitly).
-  let matchedCase = 'default';
-  for (const [caseName, caseValue] of Object.entries(config.cases)) {
-    if (value == caseValue) {
-      matchedCase = caseName;
-      break;
-    }
-  }
-
-  return { ok: true, output: { case: matchedCase, value } };
-}
-
-// -- agent-invoke dispatch --
-
-interface AgentInvokeResult {
-  ok: boolean;
-  result?: string;
-  output?: Record<string, unknown>;
-  error?: string;
-}
-
-/**
- * Execute an `agent-invoke` node. Resolves the sub-agent from the
- * AgentStore, maps inputs from upstream outputs, then recursively calls
- * `executeAgentDag`. The sub-agent gets its own `runs` row linked to
- * the parent via `parentRunId` + `parentNodeId`.
- */
-async function executeAgentInvokeNode(
-  node: AgentNode,
-  outputs: Map<string, NodeOutput>,
-  parentRunId: string,
-  parentOptions: DagExecuteOptions,
-  deps: DagExecutorDeps,
-  parentAgent: Agent,
-): Promise<AgentInvokeResult> {
-  const config = node.agentInvokeConfig;
-  if (!config) {
-    return { ok: false, error: 'agent-invoke node missing agentInvokeConfig' };
-  }
-  if (!deps.agentStore) {
-    return { ok: false, error: 'agent-invoke requires agentStore on DagExecutorDeps' };
-  }
-
-  const subAgent = deps.agentStore.getAgent(config.agentId);
-  if (!subAgent) {
-    return { ok: false, error: `Sub-agent "${config.agentId}" not found in store` };
-  }
-
-  // Map inputs from upstream outputs or the parent's caller inputs.
-  const subInputs: Record<string, string> = {};
-  if (config.inputMapping) {
-    for (const [subKey, sourceExpr] of Object.entries(config.inputMapping)) {
-      // sourceExpr can be a literal or an upstream path like "upstream.fetch.result".
-      if (sourceExpr.startsWith('upstream.')) {
-        const parts = sourceExpr.split('.');
-        const upId = parts[1];
-        const field = parts.slice(2).join('.');
-        const upOutput = outputs.get(upId);
-        if (upOutput?.outputs) {
-          const val = walkPath(upOutput.outputs, field);
-          subInputs[subKey] = val !== undefined ? String(val) : '';
-        } else if (upOutput) {
-          subInputs[subKey] = upOutput.result;
-        }
-      } else {
-        // Literal value or parent input ref.
-        subInputs[subKey] = sourceExpr;
-      }
-    }
-  }
-
-  // Recursive execution with nested-run linking.
-  const subRun = await executeAgentDag(
-    subAgent,
-    {
-      triggeredBy: parentOptions.triggeredBy,
-      inputs: subInputs,
-      parentRunId,
-      parentNodeId: node.id,
-    },
-    deps,
-  );
-
-  if (subRun.status === 'completed') {
-    // Parse the sub-run's result as structured output if possible.
-    let output: Record<string, unknown> = { result: subRun.result ?? '' };
-    if (subRun.result) {
-      try {
-        const parsed = JSON.parse(subRun.result);
-        if (typeof parsed === 'object' && parsed !== null) {
-          output = { ...parsed, result: subRun.result };
-        }
-      } catch { /* plain string result */ }
-    }
-    return { ok: true, result: subRun.result ?? '', output };
-  }
-
-  return {
-    ok: false,
-    error: `Sub-agent "${config.agentId}" failed: ${subRun.error ?? subRun.status}`,
-  };
-}
-
-// -- loop dispatch --
-
-/**
- * Execute a `loop` node. Iterates over an array extracted from upstream
- * structured output, invoking a sub-agent per item. Each iteration is
- * a nested run linked to the parent. Results are collected into an
- * `{ items: result[], count: number }` output.
- *
- * If any iteration fails, the loop continues (best-effort) and the
- * failed item is recorded as `null` in the items array. The loop node
- * itself only fails if the config is invalid or the sub-agent can't
- * be found.
- */
-async function executeLoopNode(
-  node: AgentNode,
-  outputs: Map<string, NodeOutput>,
-  parentRunId: string,
-  parentOptions: DagExecuteOptions,
-  deps: DagExecutorDeps,
-  parentAgent: Agent,
-): Promise<AgentInvokeResult> {
-  const config = node.loopConfig;
-  if (!config) {
-    return { ok: false, error: 'Loop node missing loopConfig' };
-  }
-  if (!deps.agentStore) {
-    return { ok: false, error: 'Loop node requires agentStore on DagExecutorDeps' };
-  }
-
-  const subAgent = deps.agentStore.getAgent(config.agentId);
-  if (!subAgent) {
-    return { ok: false, error: `Loop sub-agent "${config.agentId}" not found in store` };
-  }
-
-  // Extract the array to iterate over from the first upstream's output.
-  const upstreamId = (node.dependsOn ?? [])[0];
-  if (!upstreamId) {
-    return { ok: false, error: 'Loop node has no dependsOn' };
-  }
-  const upOutput = outputs.get(upstreamId);
-  if (!upOutput) {
-    return { ok: false, error: `Upstream "${upstreamId}" has no output` };
-  }
-
-  let items: unknown[];
-  if (upOutput.outputs) {
-    const arr = walkPath(upOutput.outputs, config.over);
-    if (!Array.isArray(arr)) {
-      return { ok: false, error: `Loop field "${config.over}" on upstream "${upstreamId}" is not an array` };
-    }
-    items = arr;
-  } else {
-    try {
-      const parsed = JSON.parse(upOutput.result);
-      const arr = walkPath(parsed, config.over);
-      if (!Array.isArray(arr)) {
-        return { ok: false, error: `Loop field "${config.over}" on upstream "${upstreamId}" is not an array` };
-      }
-      items = arr;
-    } catch {
-      return { ok: false, error: `Cannot parse upstream "${upstreamId}" result as JSON for loop` };
-    }
-  }
-
-  const maxIter = config.maxIterations ?? 1000;
-  const limited = items.slice(0, maxIter);
-  const results: (string | null)[] = [];
-
-  for (let i = 0; i < limited.length; i++) {
-    const item = limited[i];
-    // Each iteration passes the current item as the ITEM input to the sub-agent.
-    const subInputs: Record<string, string> = {
-      ITEM: typeof item === 'string' ? item : JSON.stringify(item),
-      ITEM_INDEX: String(i),
-    };
-
-    const subRun = await executeAgentDag(
-      subAgent,
-      {
-        triggeredBy: parentOptions.triggeredBy,
-        inputs: subInputs,
-        parentRunId,
-        parentNodeId: node.id,
-      },
-      deps,
-    );
-
-    if (subRun.status === 'completed') {
-      results.push(subRun.result ?? null);
-    } else {
-      results.push(null);
-    }
-  }
-
-  return {
-    ok: true,
-    result: JSON.stringify({ items: results, count: results.length }),
-    output: { items: results, count: results.length },
-  };
-}
-
-// -- onlyIf evaluation --
-
-/**
- * Evaluate an `onlyIf` predicate against an upstream node's output.
- * Returns true if the condition is met (node should execute), false
- * if not (node should be skipped with condition_not_met).
- *
- * The `field` is a dot-separated path into the upstream's structured
- * output (or the flat `result` string for v0.15 nodes).
- */
-function evaluateOnlyIf(condition: OnlyIfCondition, upOutput: NodeOutput | undefined): boolean {
-  if (!upOutput) return false;
-
-  // Walk the field path into the structured output if available,
-  // falling back to the flat result string.
-  let value: unknown;
-  if (upOutput.outputs) {
-    value = walkPath(upOutput.outputs, condition.field);
-  } else {
-    // v0.15 node — only "result" is available as a field.
-    if (condition.field === 'result') {
-      value = upOutput.result;
-    } else {
-      // Try parsing result as JSON and walking the path.
-      try {
-        const parsed = JSON.parse(upOutput.result);
-        value = walkPath(parsed, condition.field);
-      } catch {
-        return false;
-      }
-    }
-  }
-
-  if (condition.exists !== undefined) {
-    return condition.exists ? value !== undefined && value !== null : value === undefined || value === null;
-  }
-  if (condition.equals !== undefined) {
-    return value == condition.equals; // loose equality for string/number coercion
-  }
-  if (condition.notEquals !== undefined) {
-    return value != condition.notEquals;
-  }
-  // No predicate specified — just check the field exists.
-  return value !== undefined && value !== null;
-}
-
-/** Walk a dot-separated path into a nested object. */
-function walkPath(obj: unknown, path: string): unknown {
-  let current: unknown = obj;
-  for (const segment of path.split('.')) {
-    if (current === null || current === undefined) return undefined;
-    if (typeof current === 'object') {
-      current = (current as Record<string, unknown>)[segment];
-    } else {
-      return undefined;
-    }
-  }
-  return current;
-}
-
-const MINIMAL_ALLOWLIST = ['PATH', 'HOME', 'LANG', 'TERM', 'TMPDIR'];
-const LOCAL_ALLOWLIST = [...MINIMAL_ALLOWLIST, 'USER', 'SHELL', 'NODE_ENV', 'TZ'];
-const LOCAL_PATTERNS = [/^LC_/];
-
-// -- Tool dispatch helpers --
-
-/**
- * Derive the tool id for a node. Only nodes that explicitly set `tool:`
- * go through the tool dispatch path. v0.15 nodes without `tool:` use
- * the existing spawn path directly — desugaring to `shell-exec` /
- * `claude-code` built-ins happens at a higher layer (YAML parser /
- * import) when the user opts in, not silently at exec time.
- */
-function resolveToolId(node: AgentNode): string | undefined {
-  return node.tool;
-}
-
-/**
- * Build the inputs map for a tool invocation. For v0.16 nodes with
- * `toolInputs:`, use those directly. For v0.15 nodes, fold the inline
- * `command` / `prompt` into the shape the built-in tool expects.
- */
-function resolveToolInputs(
-  node: AgentNode,
-  upstreamSnapshot: Record<string, string>,
-): Record<string, unknown> {
-  if (node.toolInputs) return { ...node.toolInputs };
-  // Backcompat: v0.15 inline fields → built-in tool input shape.
-  if (node.type === 'shell' && node.command) {
-    return { command: node.command };
-  }
-  if (node.type === 'claude-code' && node.prompt) {
-    return {
-      prompt: node.prompt,
-      model: node.model,
-      maxTurns: node.maxTurns,
-      allowedTools: node.allowedTools,
-    };
-  }
-  return {};
 }
